+++ core/model/application.go
@@ -306,6 +306,8 @@ type IdentityProviderSetting struct {
 
 	UserSpecificFields []string `bson:"user_specific_fields"`
 
+	AlwaysSyncProfile bool `bson:"always_sync_profile"` // if true, profile data will be overwritten with data from external user on each login/refresh
+
 	Roles  map[string]string `bson:"roles"`  //map[identity_provider_role]app_role_id
 	Groups map[string]string `bson:"groups"` //map[identity_provider_group]app_group_id
 }
@@ -356,7 +358,7 @@ type AuthTypesSupport struct {
 	SupportedAuthTypes []SupportedAuthType `bson:"supported_auth_types"`
 }
 
-// AuthTypesSupport represents a supported auth type
+// SupportedAuthType represents a supported auth type
 type SupportedAuthType struct {
 	AuthTypeID string                 `bson:"auth_type_id"`
 	Params     map[string]interface{} `bson:"params"`